Add Story Reorganizer-object
Introduced to remove code duplication of CLI-module and coming
test-module
